Refactor measure_time function to ensure return type is float

+++ python_async_function/2-measure_runtime.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 
-import asyncio
 import time
 
 wait_n = __import__('1-concurrent_coroutines').wait_n
@@ -11,4 +10,4 @@ async def measure_time(n: int, max_delay: float) -> float:
     await wait_n(n, max_delay)
     end = time.time()
 
-    return (end - start)/n
+    return float((end - start)/n)
